+++ docs/resources/changelog.md
@@ -13,6 +13,7 @@ _During the beta period, these restrictions may be relaxed in the event of a mis
 - Fixed a bug where updating a menu item's label wouldn't update the display label in `<sl-select>` [#729](https://github.com/shoelace-style/shoelace/issues/729)
 - Improved performance of `<sl-select>` by caching menu items instead of traversing for them each time
 - Revert form submit logic [#718](https://github.com/shoelace-style/shoelace/issues/718)
+- Updated the `name` and `icon` attribute so they reflect in `<sl-icon>` [#742](https://github.com/shoelace-style/shoelace/pull/742)
 
 ## 2.0.0-beta.73